feat: implement protected routes
