+++ src/app/hooks.ts
@@ -3,3 +3,6 @@ import { AppDispatch, RootState } from "./store";
 
 export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
 export const useAppSelector = useSelector.withTypes<RootState>();
+export const useIsAuthenticated = () => {
+  return useAppSelector((state) => Boolean(state.auth.token));
+};
