+++ src/app/navigation/AppNavigator.tsx
@@ -1,0 +1,32 @@
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import { NavigationContainer } from "@react-navigation/native";
+import { useIsAuthenticated } from "../hooks";
+import Login from "@/src/features/auth/screens/Login";
+import Products from "@/src/features/products/Products";
+
+const AppNavigator = () => {
+  const isAuthenticated = useIsAuthenticated();
+  const Stack = createNativeStackNavigator();
+
+  return (
+    <NavigationContainer>
+      <Stack.Navigator>
+        {isAuthenticated ? (
+          <>
+            <Stack.Screen name="Products" component={Products} />
+          </>
+        ) : (
+          <>
+            <Stack.Screen
+              name="Login"
+              component={Login}
+              options={{ headerShown: false }}
+            />
+          </>
+        )}
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+};
+
+export default AppNavigator;
