+++ src/features/products/Products.tsx
@@ -1,0 +1,11 @@
+import { View, Text } from "react-native";
+
+const Products = () => {
+  return (
+    <View>
+      <Text>Products</Text>
+    </View>
+  );
+};
+
+export default Products;
